#3 - Untangled ethers from the chain module. Chain module is now completely agnostic.
