--- packages/chain/src/api/erc20.ts
@@ -1,7 +1,0 @@
-import { OiEthereumApi } from "./api";
-import { useContractAPI } from "./apis";
-
-
-export class OiErc20Api extends OiEthereumApi {
-  
-}
