+++ packages/chain/src/connectors/index.ts
@@ -1,5 +1,2 @@
-import './erc20';
-import './erc721';
-
-export { OiContractConnector as OiConnector } from "./connector";
+export { OiContractConnector, OiContractConnectorParams } from "./connector";
 export { useContractConnector, getContractConnector } from "./connectors";
