+++ packages/ethereum/src/erc1155/abi.ts
