+++ packages/ethereum/src/erc20/abi.ts
@@ -1,5 +1,3 @@
-import { useABI } from "../contracts";
-
 export const ERC20abi = [
   {
     constant: true,
