+++ packages/ethereum/src/erc20/api.ts
@@ -1,0 +1,9 @@
+import { OiEthereumApi } from "../api";
+import { useContractAPI } from "@openibex/chain";
+
+
+export class OiErc20Api extends OiEthereumApi {
+  
+}
+
+useContractAPI("erc20", OiErc20Api)
