+++ packages/ethereum/src/erc20/index.ts
@@ -1,0 +1,1 @@
+export {OiErc20} from './connector';
