+++ packages/ethereum/src/erc721/index.ts
@@ -1,0 +1,1 @@
+export {OiErc721} from './connector';
