+++ packages/ethereum/src/index.ts
@@ -1,6 +1,10 @@
 import '@openibex/chain';
 
-export {EthersProviderFactory} from './provider';
+export {EthereumProviderFactory} from './provider';
 export {EthereumSubscriberFactory} from './subscriber';
+export {EthereumContractFactory} from './contract';
 export {EthereumEventIndexer} from './indexer';
 
+import './erc20';
+import './erc721';
+import './erc1155';
